members の算出場所を修正
